+++ algorithms/degree_function.py
